Move multi_tap_listener
