+++ packages/flame/lib/src/events/flame_game_mixins/has_tappables.dart
@@ -4,7 +4,7 @@ import 'package:meta/meta.dart';
 import '../../components/mixins/tappable.dart';
 import '../../game/flame_game.dart';
 import '../../gestures/events.dart';
-import '../multi_tap_listener.dart';
+import '../interfaces/multi_tap_listener.dart';
 
 mixin HasTappables on FlameGame implements MultiTapListener {
   @mustCallSuper
